requirements.txt
faker
pandas
sqlalchemy
pymysql
pymongo
python-dotenv
tqdm
numpy
mysql
requests
kpler-sdk
load_dotenv
mysql-connector-python
websockets
flask
kafka-python
flask_socketio
websocket-client
hdfs
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
selenium>=4.15.0
